TALLER_02/Legalia/backend/requirements.txt
flask
pymysql
python-dotenv
flask-cors
flask-limiter
flask-talisman
flasgger
argon2-cffi
pyjwt
uuid6
